+++ src/lib.rs
@@ -1,17 +1,15 @@
 #![no_std]
 
 mod backend;
+mod drawtarget;
 mod font;
+mod texttarget;
 
-use core::fmt::Error as FmtError;
-use core::fmt::Result as FmtResult;
-use core::fmt::Write;
-
+use embedded_graphics_core::{pixelcolor::BinaryColor, prelude::*};
 use embedded_hal::blocking;
 use embedded_hal::digital::v2::OutputPin;
 
 use crate::backend::*;
-use crate::font::*;
 
 pub const WIDTH: u8 = 84;
 pub const HEIGHT: u8 = 48;
@@ -59,6 +57,7 @@ where
     extended_instruction_set: bool,
     x: u8,
     y: u8,
+    framebuffer: [[u8; WIDTH as usize]; ROWS as usize],
 }
 
 impl<CLK, DIN, DC, CE, LIGHT, RST, ERR> PCD8544<PCD8544GpioBackend<CLK, DIN, DC, CE>, LIGHT, RST>
@@ -93,6 +92,7 @@ where
             extended_instruction_set: false,
             x: 0,
             y: 0,
+            framebuffer: [[0; WIDTH as usize]; ROWS as usize],
         })
     }
 }
@@ -118,6 +118,7 @@ where
             extended_instruction_set: false,
             x: 0,
             y: 0,
+            framebuffer: [[0; WIDTH as usize]; ROWS as usize],
         })
     }
 }
@@ -130,8 +131,8 @@ where
 {
     pub fn reset(&mut self) -> Result<(), Backend::Error> {
         self.rst.set_low()?;
-        self.x = 0;
-        self.y = 0;
+        self.reset_position();
+        <Self as DrawTarget>::clear(self, BinaryColor::Off)?;
         self.init()
     }
 
@@ -158,9 +159,8 @@ where
     }
 
     pub fn clear(&mut self) -> Result<(), Backend::Error> {
-        for _ in 0..(WIDTH as u16 * ROWS as u16) {
-            self.write_data(0x00)?;
-        }
+        <Self as DrawTarget>::clear(self, BinaryColor::Off)?;
+        self.flush()?;
         self.set_x_position(0)?;
         self.set_y_position(0)
     }
@@ -173,26 +173,6 @@ where
     pub fn set_entry_mode(&mut self, entry_mode: bool) -> Result<(), Backend::Error> {
         self.entry_mode = entry_mode;
         self.write_current_function_set()
-    }
-
-    pub fn x(&self) -> u8 {
-        self.x
-    }
-
-    pub fn y(&self) -> u8 {
-        self.y
-    }
-
-    pub fn set_x_position(&mut self, x: u8) -> Result<(), Backend::Error> {
-        let x = x % WIDTH;
-        self.x = x;
-        self.write_command(0x80 | x)
-    }
-
-    pub fn set_y_position(&mut self, y: u8) -> Result<(), Backend::Error> {
-        let y = y % ROWS;
-        self.y = y;
-        self.write_command(0x40 | y)
     }
 
     pub fn set_light(&mut self, enabled: bool) -> Result<(), Backend::Error> {
@@ -259,44 +239,6 @@ where
     }
 
     fn write_data(&mut self, value: u8) -> Result<(), Backend::Error> {
-        self.increase_position();
-
         self.backend.write_byte(true, value)
-    }
-
-    fn increase_position(&mut self) {
-        self.x = (self.x + 1) % WIDTH;
-        if self.x == 0 {
-            self.y = (self.y + 1) & ROWS;
-        }
-    }
-}
-
-impl<Backend, LIGHT, RST> Write for PCD8544<Backend, LIGHT, RST>
-where
-    Backend: PCD8544Backend,
-    LIGHT: OutputPin<Error = Backend::Error>,
-    RST: OutputPin<Error = Backend::Error>,
-{
-    fn write_str(&mut self, s: &str) -> FmtResult {
-        for char in s.chars() {
-            match char {
-                '\r' => {
-                    self.set_x_position(0).map_err(|_| FmtError)?;
-                }
-                '\n' => {
-                    for _ in 0..(WIDTH - self.x) {
-                        self.write_data(0x00).map_err(|_| FmtError)?;
-                    }
-                }
-                _ => {
-                    for b in char_to_bytes(char) {
-                        self.write_data(*b).map_err(|_| FmtError)?;
-                    }
-                    self.write_data(0x00).map_err(|_| FmtError)?;
-                }
-            }
-        }
-        Ok(())
     }
 }